feat: add command to list sessions
Closes SHIP-2639
